Move CaptureSession Protocols to the new file
